% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLARMA.R
\name{fitted.GLARMA}
\alias{fitted.GLARMA}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{GLARMA}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values.
}
\examples{
tsibbledata::aus_production |>
  fabletools::model(manual_gla = GLARMA(Beer ~ pq(1,0))) |>
  dplyr::select(manual_gla) |>
  fitted()
}
