% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INGARCH.R
\name{fitted.INGARCH}
\alias{fitted.INGARCH}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{INGARCH}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values.
}
\examples{
tsibbledata::aus_production |>
  fabletools::model(manual_ing = INGARCH(Beer ~ pq(1,1))) |>
  dplyr::select(manual_ing) |>
  fitted()

}
