% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLARMA.R
\name{residuals.GLARMA}
\alias{residuals.GLARMA}
\title{Extract residuals from a fable model}
\usage{
\method{residuals}{GLARMA}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of fitted residuals.
}
\description{
Extracts the residuals.
}
\examples{
tsibbledata::aus_production |>
  fabletools::model(manual_gla = GLARMA(Beer ~ pq(1,0))) |>
  dplyr::select(manual_gla) |>
  residuals()
}
