% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INGARCH.R
\name{residuals.INGARCH}
\alias{residuals.INGARCH}
\title{Extract residuals from a fable model}
\usage{
\method{residuals}{INGARCH}(object, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{...}{Other arguments passed to methods}
}
\value{
A vector of fitted residuals.
}
\description{
Extracts the residuals.
}
\examples{
tsibbledata::aus_production |>
  fabletools::model(manual_ing = INGARCH(Beer ~ pq(1,1) + PQ(1,1))) |>
  dplyr::select(manual_ing) |>
  residuals()

}
