% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLARMA.R
\name{tidy.GLARMA}
\alias{tidy.GLARMA}
\title{Tidy a fable model}
\usage{
\method{tidy}{GLARMA}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments to tidying method.}
}
\value{
The model's coefficients in a `tibble`.
}
\description{
Returns the coefficients from the model in a `tibble` format.
}
\examples{
tsibbledata::aus_production |>
  fabletools::model(manual_gla = GLARMA(Beer ~ pq(1,0))) |>
  dplyr::select(manual_gla) |>
  fabletools::tidy()

}
