% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INGARCH.R
\name{tidy.INGARCH}
\alias{tidy.INGARCH}
\title{Tidy a fable model}
\usage{
\method{tidy}{INGARCH}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments to tidying method.}
}
\value{
The model's coefficients in a `tibble`.
}
\description{
Returns the coefficients from the model in a `tibble` format.
}
\examples{
tsibbledata::aus_production |>
  fabletools::model(manual_ing = INGARCH(Beer ~ pq(1,1))) |>
  dplyr::select(manual_ing) |>
  fabletools::tidy()

}
